requirements.txt
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
distro==1.9.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
Flask==3.1.3
flask-cors==6.0.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ifcopenshell==0.8.4.post1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
lark==1.3.1
MarkupSafe==3.0.3
neo4j==6.1.0
numpy==2.4.3
openpyxl==3.1.5
pandas==3.0.1
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
reportlab==4.4.10
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
Werkzeug==3.1.6
